Add versioned evaluations with tags and metadata
