+++ bin/eval-weather-bot.js
@@ -1,5 +1,7 @@
 import { ChatModelService, ChatModelType } from '../src/ai/chat-model.js';
 import { loadDataset, runEvaluation } from '../src/eval.js';
+import WeatherAgentV1 from '../src/ai/agents/weather-agent/v1.js';
+import WeatherAgentV2 from '../src/ai/agents/weather-agent/v2.js';
 import dotenv from 'dotenv';
 
 // load .env
@@ -24,54 +26,9 @@ const maxTokens = 2000;
 const model = ChatModelType.GPT_4O_MINI;
 const service = ChatModelService.OPENAI;
 
-const WeatherAgent = {
-	id: 'weatherbot',
-	name: 'Weather Bot',
-	instructions: 'You are a helpful weather bot',
-	additionalInstructions: ( context ) => {
-		return `The user's current location is ${ context.currentLocation }.`;
-	},
-	toolkits: [
-		{
-			name: 'weather',
-			context: {
-				currentLocation: 'Melbourne, Australia',
-			},
-			tools: [
-				{
-					name: 'getWeather',
-					description: 'Get the weather for a location',
-					parameters: {
-						type: 'object',
-						properties: {
-							location: {
-								type: 'string',
-							},
-						},
-						required: [ 'location' ],
-					},
-				},
-			],
-			callbacks: {
-				getWeather: async ( { location } ) => {
-					const response = await fetch(
-						`https://wttr.in/${ location }?format=%C+%t`
-					);
-					const text = await response.text();
-					return text;
-				},
-			},
-		},
-	],
-	onStart: ( invoke ) => {
-		console.warn( 'Weather Agent started' );
-		invoke.agentSay( 'What location would you like the weather for?' );
-	},
-};
-
 const result = await runEvaluation(
 	'Big Sky',
-	[ WeatherAgent ],
+	[ WeatherAgentV1, WeatherAgentV2 ],
 	dataset,
 	model,
 	service,
